First draft of dev OSM host setup than mostly works.

+++ orchestrator/env/setup.sh
@@ -2,21 +2,26 @@
 
 [ $(id -u) == 0 ] || exec sudo -- "$0" "$@"
 
-# debconf-get-selections --installer"
+if [ -z "$(which qemu-system-x86_64)" ]
+then
+   echo "Press install qemu-system-x86"
+   exit -1
+fi
+
+if [ -e "/usr/share/OVMF/OVMF_CODE_4M.fd" ]
+then
+   echo "Press install ovmf"
+   exit -1
+fi
 
 DEBISO=debian-12.5.0-amd64-netinst.iso
 DEBDISK=disk.qcow
-DEBBIOSMEM=ovmf_vars.fd
 
 mkdir -p mnt
 mount $DEBISO mnt
 mkdir -p boot
 cp -r mnt/install.amd boot/
 umount mnt
-
-# apt install ovmf qemu-system-x86
-
-dd if=/dev/zero of="$DEBBIOSMEM" bs=131072 count=1
 
 qemu-img create -f qcow2 "$DEBDISK" 16G
 
@@ -39,7 +44,6 @@ qemu-system-x86_64                 \
    -drive file="$DEBISO",format=raw,if=virtio,media=cdrom \
    -drive file="$DEBDISK",format=qcow2,if=virtio \
    -drive "if=pflash,format=raw,unit=0,file=/usr/share/OVMF/OVMF_CODE_4M.fd,readonly=on" \
-   -drive "if=pflash,format=raw,unit=1,file=$DEBBIOSMEM" \
    -kernel boot/install.amd/vmlinuz \
    -initrd boot/install.amd/initrd.gz \
    -append "console=ttyS0 priority=critical auto=true DEBIAN_FRONTEND=text log_host=$IP_ADDR log_port=10514 url=http://$IP_ADDR:8000/preseed.cfg"
